renamed dummy page and links to dashboard
